Add BrerEventListenerWithHandlersMock and enhance tests
This commit extends the tests for the Brer listener by introducing `BrerEventListenerWithHandlersMock` and `ExtraBrerEventListenerWithHandlersMock`. It helps to test subscriber methods and attribute events. It also adds assertions to make sure that the listener class has correctly registered the event and handlers. The `BrerEventListenerWithoutHandlersMock` has been moved to the Helpers directory for better organization.