Down the rabbit hole! Figured out a way to actually test this. Add exception for invalid BrerHandler parameters count
This commit introduces a new exception - InvalidBrerHandlerParameterCountException, which is thrown when an invalid number of parameters are provided to the BrerHandler. The number of parameters should be exactly one. Any deviation from this requirement (i.e., when no parameter or more than one parameter is provided) is considered invalid and triggers this exception. Corresponding tests are also added to assert the correctness of this functionality.